==> 04_Flow_Data_Shiny.Rmd <==
---
title: "Notebook 04: Using Shiny App to wrap up EDA in notebok 03"
runtime: shiny
output: html_document
---

# Goal:
This notebook contains the shiny app that includes the exploratory data analysis in notebook 03. The app allows user to select the fcs sample and also to customize the parameters of analysis and visualization. (**Caveat**) Since the dataset is huge, for some parameters setting, the plot may require some time to generate.

# Setup environment

import needed library
```{r, results="hide", warning=FALSE, message=FALSE}
# utilities
library(tidyverse)
library(Rtsne)
library(reshape2)

# visualization
library(shiny)
library(pheatmap)
library(networkD3)
library(RColorBrewer)
library(viridis)

# flow cytometry
library(flowCore)
library(flowViz)
library(ggcyto)
```

setup required directory
```{r}
datadir <- "/data/Flow_HIV"
```

# Set up Shiny Applications

**Global**
```{r, results="hide", warning=FALSE, message=FALSE}
#### Load data ####
data(MisLinks)
data(MisNodes)
dat_clinic <- read_delim(
    file.path(datadir, "attachments", "clinical_data_flow_repository.csv"),
    delim = "\t")

#load(file.path(datadir, "FlowHIV_Preprocessed.RData"))
load(file.path(datadir, "FlowHIV_Preprocessed_subset10.RData"))

# set parameters
fclist  <- fs_trans_subset10@phenoData@data$name
set_seed <- 0

fdat <- fs_trans_subset10[[1]]
tmp1 <- fdat@parameters@data$name
tmp2 <- fdat@parameters@data$desc
len  <- length(tmp1)
col_name = c(
    tmp1[1:3], 
    paste(tmp1[4:len], tmp2[4:len]))
names(col_name) <- NULL
```

**Server**
```{r}
#### Server ####
server <- function(input, output) {
    
    output$summary <- renderPlot({
        dat <- dat_clinic
        dat <- dat %>% arrange(`Survival time`)
        dat$ID <- as.character(dat$ID)  
        dat$ID <- factor(dat$ID, levels = dat$ID)
        
        # xxx.fcs -> xxx
        id <- str_split(input$file_name, "\\.", simplify = TRUE)[1,1]

        dat$choose <- ifelse(dat$ID %in% id, "Chosen", "Not_Chosen")
        loc <- which(dat$choose == "Chosen")

        gp <- ggplot(dat, aes(
                x = ID, 
                y = `Survival time`,
                fill = choose)) + 
            geom_histogram(stat = "identity") +
            scale_fill_manual(
                "legend", 
                values = c("Chosen" = "#fbb4ae", "Not_Chosen" = "grey50")) +
            geom_vline(xintercept = loc, color = "red") +
        theme(axis.text.x = element_text(size = 2, angle = 90, hjust = 1))
        
        print(gp)
    })
    
    output$distribution <- renderPlot({
        fdat <- fs_trans_subset10[[input$file_name]]
        autoplot(fdat)
    }) # end render summary
    
    output$pairwise <- renderPlot({
        fdat <- fs_trans_subset10[[input$file_name]]
        print(splom(fdat))
    }) # end render scatter
    
    output$scatter <- renderPlot({
        fdat <- fs_trans_subset10[[input$file_name]]
        
        # input$marker1, input$marker2 -> ex: "B515-A KI67"
        # what we want: B515-A
        marker1 <- str_split(input$marker1, " ", simplify = TRUE)[1,1]
        marker2 <- str_split(input$marker2, " ", simplify = TRUE)[1,1]
        
        # scatter plot
        if (marker1 == marker2){
            autoplot(fdat, marker1)
        } else {
            autoplot(fdat, marker1, marker2)
        } # end if-else
        
        
    }) # end render scatter

    output$heatmap <- renderPlot({
        fdat <- fs_trans_subset10[[input$file_name]]
        
        expr <- data.frame(exprs(fdat))
        colnames(expr) <- col_name
        
        set.seed(set_seed)
        idx_events <- sample(
            1:nrow(expr), 
            floor(input$fraction * nrow(expr)))
        
        mat <- expr[idx_events, 4:16]
        
        # colors (ex: color = inferno(10))
        color = brewer.pal(9, name = "Reds")
        color = c(
            rep(color[1:3], each = 5), 
            color[3:length(color)])
       
        
        pheatmap(
            mat               = mat,
            color             = color,
            border_color      = NA,
            show_colnames     = TRUE, #FALSE,
            show_rownames     = FALSE,
            #annotation_col    = mat_col,
            #annotation_colors = mat_colors,
            drop_levels       = TRUE,
            fontsize          = 14,
            main              = paste("Heatmap", input$file_name))
    }) # end render heatmap
    
    output$tsne_plot <- renderPlot({
        fdat <- fs_trans_subset10[[input$file_name]]
        expr <- data.frame(exprs(fdat))

        set.seed(set_seed)
        idx_events <- sample(
            1:nrow(expr), 
            floor(input$fraction * nrow(expr)))
        mat <- expr[idx_events, ]
        
        tsne_out = Rtsne(
            mat,
            perplexity = input$perplexity,
            theta = input$theta,
            max_iter = input$max_iter)
        
        plot(
            tsne_out$Y, pch = 20,
            xlab = "t-sne axis x",
            ylab = "t-sne axis y")
    }) # end render tsne plot
    
    output$simple <- renderSimpleNetwork({
        fdat <- fs_trans_subset10[[input$file_name]]
        expr <- exprs(fdat)
        expr <- apply(expr, 2, scale)
        
        set.seed(set_seed)
        idx_events <- sample(
            1:nrow(expr), 
            floor(input$fraction * nrow(expr)))
        
        mat <- expr[idx_events, ]
        mat_dist <- dist(mat)
        mat_sim  <- 1 / (as.matrix(mat_dist) + 1)
        diag(mat_sim) <- 0
    
        cutoff <- input$cutoff
        adj   <- ifelse(mat_sim > cutoff, 1, 0)
        links <- melt(adj) %>% 
            dplyr::filter(value == 1) %>% 
            dplyr::select(-value)
        
        simpleNetwork(
            links,
            fontSize     = input$fontSize,
            charge       = input$charge,
            linkDistance = input$linkDistance,
            zoom         = TRUE
        )
    }) # end render SimpleNetwork  
    
} # end server


```

**User Interface**
```{r}
#### UI ####
ui <- shinyUI(fluidPage(
    # title
    titlePanel("Explore Flow HIV Data"),
    
    # sideLayout
    sidebarLayout(
        sidebarPanel(
            helpText("Data Selection"),
            selectInput("file_name", "Select Your Data", fclist, 
                  selected = "100715.fcs", multiple = FALSE, selectize = TRUE, 
                  width = NULL, size = NULL),
            
            helpText("Scatter Plot"),
            selectInput("marker1", "Select Marker 1 (X-axis)", 
                  col_name,
                  #colnames(fs_trans_subset10), 
                  selected = "FSC-A", multiple = FALSE, selectize = TRUE, 
                  width = NULL, size = NULL),
            selectInput("marker2", "Select Marker 2 (Y-axis)", 
                  col_name,
                  #colnames(fs_trans_subset10), 
                  selected = "FSC-H", multiple = FALSE, selectize = TRUE, 
                  width = NULL, size = NULL),
            
            helpText("Fraction (Used in Heatmap, t-SNE, and network)"),
            sliderInput(
                "fraction", "Fraction of Events", 0.01, 
                min = 0.01, max = 0.5, step = .01),
            
            helpText("T-SNE Plot"),
            sliderInput(
                "perplexity", "Perplexity", 30, 
                min = 1, max = 100, step = 10),
            sliderInput(
                "theta", "Theta", 0.5, 
                min = 0, max = 1.0, step = .1),
            sliderInput(
                "max_iter", "Max Iteration", 500, 
                min = 100, max = 2000, step = 100),
            
            helpText("Sample Network Plot"),
            sliderInput(
                "cutoff", "Cut Off Value", 0.7, 
                min = 0.5, max = 0.9, step = .01),
            sliderInput("charge", "Charge", -30, 
                  min = -100, max = 0, step = 10),
            sliderInput("linkDistance", "Sample Network: Link Distance", 5, 
                  min = 0.01, max = 10, step = 1),
            sliderInput("fontSize", "Font Size", 15, 
                  min = 5, max = 50, step = 5)
            ),
        mainPanel(
            tabsetPanel(
                tabPanel("Summary",
                         plotOutput("summary"),
                         #plotOutput("pairwise"), 
                         plotOutput("distribution")),
                tabPanel("Scatter Plot",   plotOutput("scatter")),
                tabPanel("Heatmap",        plotOutput("heatmap")),
                tabPanel("t-SNE",          plotOutput("tsne_plot")),
                tabPanel("Sample Network", simpleNetworkOutput("simple"))
            ) # end tabsetPanel
        ) # end main Panel
    ) # end sidebarLayout
)) # end fluidPage & ShinyUI
```

# Run Shiny Application
```{r}
#### Run ####
shinyApp(ui = ui, server = server)
```

The results show that...
